Reactor, Bugfixes: see https://github.com/privacysandbox/noise-lab/pull/38#issuecomment-1632425375
