+++ public/index.html
@@ -83,9 +83,9 @@ limitations under the License. -->
                     <div>
                         <h3>Your conversion data</h3>
                         <div>
-                            <label for="daily">Average daily attributable conversion
+                            <label for="event-count-daily-bucket">Average daily attributable conversion
                                 count PER BUCKET:</label>
-                            <select name="daily" id="daily">
+                            <select name="event-count-daily-bucket" id="event-count-daily-bucket">
                                 <option value="5">5</option>
                                 <option value="10">10</option>
                                 <option value="100" selected>100</option>
@@ -149,7 +149,7 @@ limitations under the License. -->
                     </div>
                 </div>
                 <div class="buttons-wrapper-launch-panel">
-                    <button id="button-simulate" class="primary" onclick="simulateAndDisplayResultsSimpleMode()">
+                    <button id="button-simulate" class="primary" onclick="simulateAndDisplayResults_simpleMode()">
                         ▶️ SIMULATE
                     </button>
                 </div>
@@ -158,7 +158,7 @@ limitations under the License. -->
         <main>
             <h2>Simulations</h2>
             <div class="global-buttons">
-                <button class="primary" onclick="downloadAll_simpleMode()">
+                <button class="primary" onclick="downloadAllDom()">
                     ⬇️ Download all (.xlsx)
                 </button>
                 <button class="secondary destructive" onclick="resetUi()">
@@ -175,7 +175,7 @@ limitations under the License. -->
                 </div>
             </div>
 
-            <div class="all-simulations-wrapper" id="all-simulations-wrapper-simple-mode">
+            <div class="all-simulations-wrapper" id="all-simulations-wrapper">
             </div>
         </main>
     </div>
@@ -203,9 +203,9 @@ limitations under the License. -->
                     </div>
                     <div>
                         <h3>Your conversion data</h3>
-                        <label for="event-count">TOTAL daily attributable conversion
+                        <label for="event-count-daily-total">TOTAL daily attributable conversion
                             count:</label>
-                        <select name="event-count" id="event-count" oninput="updateDailyPerBucket()">
+                        <select name="event-count-daily-total" id="event-count-daily-total" oninput="updateDailyPerBucket()">
                             <option value="5">5</option>
                             <option value="10">10</option>
                             <option value="100">100</option>
@@ -215,9 +215,9 @@ limitations under the License. -->
                             <option value="100000">100000</option>
                         </select>
                         <div class="help" id="help-daily-total"></div>
-                        <label for="daily">Average daily attributable conversion
+                        <label for="event-count-daily-bucket">Average daily attributable conversion
                             count PER BUCKET:</label>
-                        <input name="daily" id="daily" disabled>
+                        <input name="event-count-daily-bucket" id="event-count-daily-bucket" disabled>
                         </input>
                         <div class="help" id="help-daily"></div>
                         <label for="zero-pct"> Percentage of buckets with 0 conversions: </label>
@@ -299,7 +299,6 @@ limitations under the License. -->
                                 <div id="budget-split-wrapper-el"></div>
                             </div>
                         </div>
-                        <label for="granularity">Key strategy:</label>
                         <div>
                             <label for="key-strategy">Key strategy:</label>
                             <!-- Options are populated automatically -->
@@ -321,7 +320,7 @@ limitations under the License. -->
                     </div>
                 </div>
                 <div class="buttons-wrapper-launch-panel">
-                    <button class="primary" onclick="simulateAndDisplayResultsAdvancedMode()">
+                    <button class="primary" onclick="simulateAndDisplayResults_advancedMode()">
                         ▶️ Simulate
                     </button>
                 </div>
@@ -331,7 +330,7 @@ limitations under the License. -->
             <div id="form-validation-wrapper" class="hidden"></div>
             <h2>Simulations</h2>
             <div class="global-buttons">
-                <button class="primary" onclick="downloadAll_advancedMode()">
+                <button class="primary" onclick="downloadAllDom()">
                     ⬇️ Download all (.xlsx)
                 </button>
                 <button class="secondary destructive" onclick="resetUi()">
@@ -347,7 +346,7 @@ limitations under the License. -->
                     </p>
                 </div>
             </div>
-            <div class="all-simulations-wrapper" id="all-simulations-wrapper-advanced-mode"></div>
+            <div class="all-simulations-wrapper" id="all-simulations-wrapper"></div>
         </main>
     </div>
 </body>
